refactor : formatDate 함수 수정
